Replaced DecreasingPruningHeightError with BelowPrunedThresholdError
