+++ module/mempool/incorporated_result_seals.go
@@ -37,6 +37,6 @@ type IncorporatedResultSeals interface {
 	// Monotonicity Requirement:
 	// The pruned height cannot decrease, as we cannot recover already pruned elements.
 	// If `height` is smaller than the previous value, the previous value is kept
-	// and the sentinel empool.NewDecreasingPruningHeightError is returned.
+	// and the sentinel empool.NewBelowPrunedThresholdError is returned.
 	PruneUpToHeight(height uint64) error
 }
